PipeDreams Prototype 2
Updated Pipe Dreams Prototype with functional dynamic grid generation
with random tiles, neighbor detection, and grid updating with lit
squares. Known issue is that the program is using absolute paths for the
images, so you will need to change the image loading code to get the
game to run with the proper pipes.