Final Project Submission
So, it kind of works. Basically, I got all of the logic of ships
spawning, targeting, moving, attacking, shooting, and dying to work.
What's not working is that the pathfinding gets totally screwing under
certain conditions, and I basically had to hardcode movement patterns to
make it viewable. I'll keep working on it, but this is likely as good as
it will get.